requirements.txt
jinja2>=3.1
Pillow>=10.0
